+++ R/extendr-wrappers.R
@@ -57,13 +57,19 @@ clone_robj <- function(robj) .Call(wrap__clone_robj, robj)
 
 test_robj_to_usize <- function(robj) .Call(wrap__test_robj_to_usize, robj)
 
+test_robj_to_f64 <- function(robj) .Call(wrap__test_robj_to_f64, robj)
+
 test_robj_to_i64 <- function(robj) .Call(wrap__test_robj_to_i64, robj)
 
 test_robj_to_u32 <- function(robj) .Call(wrap__test_robj_to_u32, robj)
 
+test_robj_to_i32 <- function(robj) .Call(wrap__test_robj_to_i32, robj)
+
 test_print_string <- function(s) invisible(.Call(wrap__test_print_string, s))
 
 test_robj_to_expr <- function(robj) .Call(wrap__test_robj_to_expr, robj)
+
+test_wrong_call_pl_lit <- function(robj) .Call(wrap__test_wrong_call_pl_lit, robj)
 
 test_rpolarserr <- function() .Call(wrap__test_rpolarserr)
 
@@ -927,8 +933,6 @@ LazyFrame$describe_optimized_plan <- function() .Call(wrap__LazyFrame__describe_
 
 LazyFrame$collect <- function() .Call(wrap__LazyFrame__collect, self)
 
-LazyFrame$collect_handled <- function() .Call(wrap__LazyFrame__collect_handled, self)
-
 LazyFrame$collect_in_background <- function() .Call(wrap__LazyFrame__collect_in_background, self)
 
 LazyFrame$first <- function() .Call(wrap__LazyFrame__first, self)
@@ -996,6 +1000,10 @@ LazyFrame$melt <- function(id_vars, value_vars, value_name, variable_name, strea
 LazyFrame$rename <- function(existing, new) .Call(wrap__LazyFrame__rename, self, existing, new)
 
 LazyFrame$schema <- function() .Call(wrap__LazyFrame__schema, self)
+
+LazyFrame$optimization_toggle <- function(type_coercion, predicate_pushdown, projection_pushdown, simplify_expr, slice_pushdown, cse, streaming) .Call(wrap__LazyFrame__optimization_toggle, self, type_coercion, predicate_pushdown, projection_pushdown, simplify_expr, slice_pushdown, cse, streaming)
+
+LazyFrame$profile <- function() .Call(wrap__LazyFrame__profile, self)
 
 LazyFrame$explode <- function(columns, dotdotdot_args) .Call(wrap__LazyFrame__explode, self, columns, dotdotdot_args)