+++ tests/testthat/test-dataframe.R
@@ -581,9 +581,10 @@ test_that("drop_nulls", {
   expect_equal(pl$DataFrame(tmp)$drop_nulls("mpg")$height, 29, ignore_attr = TRUE)
   expect_equal(pl$DataFrame(tmp)$drop_nulls("hp")$height, 32, ignore_attr = TRUE)
   expect_equal(pl$DataFrame(tmp)$drop_nulls(c("mpg", "hp"))$height, 29, ignore_attr = TRUE)
-  expect_grepl_error(
-    pl$DataFrame(mtcars)$drop_nulls("bad")$height,
-    "ColumnNotFound"
+
+  expect_identical(
+    result(pl$DataFrame(mtcars)$drop_nulls("bad column name")$height)$err$contexts(),
+    list(PolarsError = "not found: bad column name")
   )
 })
 
@@ -965,6 +966,13 @@ test_that("describe", {
     pl$DataFrame(mtcars)$describe(perc = numeric())$to_list(),
     pl$DataFrame(mtcars)$describe(perc = NULL)$to_list()
   )
+
+  # names using internal separator ":" in column names, should also just work.
+  df = pl$DataFrame("foo:bar:jazz" = 1, pl$Series(2, name = ""), "foobar" = 3)
+  expect_identical(
+    df$describe()$columns,
+    c("describe", df$columns)
+  )
 })
 
 test_that("glimpse", {
@@ -1019,8 +1027,8 @@ test_that("explode", {
     )
   )
 })
-  
+
 test_that("with_row_count", {
   df = pl$DataFrame(mtcars)
-  expect_identical(df$with_row_count("idx", 42)$select(pl$col("idx"))$to_data_frame()$idx, as.double(42:(41+nrow(mtcars))))
+  expect_identical(df$with_row_count("idx", 42)$select(pl$col("idx"))$to_data_frame()$idx, as.double(42:(41 + nrow(mtcars))))
 })